fix(scala): unify maxConcurrent fallback to Long (Scala 3 union compile error)
extractIntField returns Option[Long] but the AFL_MAX_CONCURRENT env fallback used
s.toInt, so orElse produced Option[Long | Int] and .map(_.toInt) failed (toInt
not a member of the union). Make the fallback s.toLong so both branches are Long.

+++ agents/scala/fw-agent/src/main/scala/afl/agent/AgentPollerConfig.scala
@@ -80,7 +80,7 @@ object AgentPollerConfig:
       .orElse(sys.env.get("AFL_POLL_INTERVAL_MS").flatMap(s => scala.util.Try(s.toLong).toOption))
       .getOrElse(2000L)
     val maxConc = extractIntField(json, "maxConcurrent")
-      .orElse(sys.env.get("AFL_MAX_CONCURRENT").flatMap(s => scala.util.Try(s.toInt).toOption))
+      .orElse(sys.env.get("AFL_MAX_CONCURRENT").flatMap(s => scala.util.Try(s.toLong).toOption))
       .map(_.toInt)
       .getOrElse(5)
     val hbMs = extractIntField(json, "heartbeatIntervalMs")
